+++ doc/release_notes.rst
@@ -29,6 +29,7 @@ Version 0.5.4
   gap tolerance.
 * Improve the mapping of termination conditions for the SCIP solver
 * Treat GLPK's `integer undefined` status as not-OK
+* Allow the objective to be set by a variable instead of an expression
 * Internally assign new data fields to `Variable` and `Constraint` with a multiindexed-safe routine. Before the
   assignment when using multi-indexed coordinates, an deprecation warning was raised. This is fixed now.